only apply to race sessions, move function out
